requirements.txt
# Core LLM and RAG dependencies
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22

# Embedding and retrieval
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4

# Reranking
cohere>=4.47

# Web interface
streamlit>=1.31.0
gradio>=4.16.0

# Observability
langfuse>=2.21.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
